# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Star Wars REST API with Swagger",
    "version": "0.1.0",
    "description": "Welcome to the swapi, the Star Wars API! This documentation should help you familiarise yourself with the resources available and how to consume them with HTTP requests.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Corentin Goujon-Cholet",
      "email": "corentingcholet@gmail.com"
    }
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "People": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birth_year": {
            "type": "string"
          },
          "eye_color": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "hair_color": {
            "type": "string"
          },
          "height": {
            "type": "string"
          },
          "mass": {
            "type": "string"
          },
          "skin_color": {
            "type": "string"
          },
          "homeworld": {
            "type": "string"
          },
          "films": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "species": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "starships": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vehicles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Film": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "episode_id": {
            "type": "number"
          },
          "opening_crawl": {
            "type": "string"
          },
          "director": {
            "type": "string"
          },
          "producer": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "species": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "starships": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vehicles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "characters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "planets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/api/film": {
      "post": {
        "tags": [
          "Film"
        ],
        "description": "Get all films",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Film"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Film"
        ],
        "description": "Get all film",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Film"
        ],
        "description": "Delete all Films",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/film/{id}": {
      "get": {
        "tags": [
          "Film"
        ],
        "description": "Get film by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The film id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "put": {
        "tags": [
          "Film"
        ],
        "description": "Modify film by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The film id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/Film"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Film"
        ],
        "description": "Delete film by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The film id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/people": {
      "post": {
        "tags": [
          "People"
        ],
        "description": "Insert people",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/People"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "get": {
        "tags": [
          "People"
        ],
        "description": "Get all peoples",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "People"
        ],
        "description": "Delete all peoples",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
    "/api/people/{id}": {
      "get": {
        "tags": [
          "People"
        ],
        "description": "Get people by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of people to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "put": {
        "tags": [
          "People"
        ],
        "description": "Modify people by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The people id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/People"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      },
      "delete": {
        "tags": [
          "People"
        ],
        "description": "Delete people by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The people id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Error"
          }
        }
      }
    },
...
  }
}